vga: FPGA tested and working!
Some quirks: The arty fpgas have a 100Mhz clock gen, so the closest
we get to vga is 25Mhz. That gives a 31.25khz line frequency and a
62.5Hz frame frequency, which is exactly what we got off of the
fpga (measured on a saleae). That made the resolution detect as
860-something x 480, which should be fine since the pixels will
stretch to match. If we had a better clock gen then we would
have better results.

The user_project_wrapper as-tested is in verilog/dv/vga.