top_level: Customasm compilation, linking, and inserting into linked binary
For testing (and part of the compilation process) the assembled binaries
are dumped to the same directory as the sources.

Shader programs are inserted directly into the finished binary and they can
be referenced from C using the linker symbols. Everything's word-aligned,
so we can word-copy directly from flash to the core IMEM.