top_level: Detect fixed point immediate
Customasm doesn't have support for floats or strings as args and can't
do conditionals in a recipe. So, both sides of the decimal point are
treated as integers. That means that leading 0s are lost on the fractional
part, and we can't count trailing 0s to figure out our fixed point conversion.

TLDR the user now has to pad out all decimals to 6 digits. I also can't set up
asserts for this because customasm isn't that good so it's your problem
to remember (this definitely won't bite me later)